renamed components and methods
