+++ csm_web/frontend/src/components/resource_aggregation/ResourceRow.tsx
@@ -1,101 +1,55 @@
-import React, { useState, useEffect } from "react";
-import ResourceEdit from "./ResourceEdit";
-import { Resource, ResourceRowProps } from "./ResourceTypes";
-import ResourceRowRender from "./ResourceRowRender";
+import React from "react";
 
-export const ResourceRow = ({ initialResource, updateResource, deleteResource, canEdit, addingResource, cancelOverride }: ResourceRowProps) => {
-  // call updateResource(resource) on change
-  const [edit, setEdit] = useState(false);
-  const [resource, setResource]: [Resource, Function] = useState({} as Resource);
-  const [fileFormData, setFileFormData] = useState(new FormData());
-
-  // update current resource if initialResource changes
-  useEffect(() => {
-    setResource(initialResource);
-    console.log(initialResource);
-    setEdit(false);
-  }, [initialResource]);
-
-  useEffect(() => {
-    setEdit(addingResource);
-  }, [addingResource]);
-
-  /**
-   * Modifies a specified field of the current resource.
-   *
-   * @param e - onChange event
-   * @param field - resource field to change
-   */
-  function handleChange(e, field) {
-    resource[field] = e.target.value;
-    setResource(resource);
-  }
-
-  /**
-   * Modifies a specified file field of the current resource.
-   *
-   * @param e - onChange event
-   * @param field - resource field to change
-   */
-  function handleFileChange(e, field) {
-    fileFormData.set(field, e.target.files[0]);
-    setFileFormData(fileFormData);
-  }
-
-  /**
-   * Bubbles change up to parent ResourceWrapper
-   *
-   * @param e - onSubmit event
-   */
-  function handleSubmit(e) {
-    e.preventDefault();
-    updateResource(resource, fileFormData);
-    setEdit(false);
-  }
-
-  /**
-   * Sets editing behavior depending on whether we can edit this resource/course
-   */
-  function handleSetEdit() {
-    if (canEdit) {
-      setEdit(true);
-    }
-  }
-
-  /**
-   * Sets edit state to false when user exits out of edit mode.
-   */
-  function handleCancel() {
-    if (cancelOverride) {
-      cancelOverride();
-    } else {
-      setEdit(false);
-    }
-  }
-
-  return (
-    <div>
-      {
-        !addingResource &&
-        <ResourceRowRender resource={resource} canEdit={canEdit} handleSetEdit={handleSetEdit} handleDelete={deleteResource} />
-      }
-      {
-        edit &&
-        <ResourceEdit
-          resource={resource}
-          handleChange={handleChange}
-          handleFileChange={handleFileChange}
-          handleSubmit={handleSubmit}
-          handleCancel={handleCancel}
-        />
-      }
+const ResourceTopics = ({ topics }) => {
+  if (topics === undefined) return <div></div>;
+  // TODO: handle multiple topics with delimiters
+  return topics.split("\x00").map((topic, index) => (
+    <div className="topic" key={index}>
+      {topic.trim()}
     </div>
-  )
+  ));
 };
 
-ResourceRow.defaultProps = {
-  addingResource: false,
-  cancelOverride: null
+const ResourceRowRender = ({ resource, canEdit, onSetEdit, onDelete }) => {
+
+  return (
+    <div className="resourceContainer">
+      <div className="resourceInfo weekNum">
+        <div>Week {resource.weekNum}</div>
+      </div>
+      <div className="resourceInfo dateCell">
+        <div>{resource.date}</div>
+      </div>
+      <div className="resourceInfo resourceTopics">
+        <div>
+          <ResourceTopics topics={resource.topics} />
+        </div>
+      </div>
+      <div className="resourceWkst">
+        <div className="resourceWkstFile">
+          <a href={resource.worksheetFile} target="_blank">
+            {/*resource.worksheetName*/}Worksheet Name
+                </a>
+        </div>
+        <div className="resourceSoln">
+          <a href={resource.solutionFile} target="_blank">
+            Solutions
+                </a>
+        </div>
+      </div>
+      {
+        canEdit &&
+        <div>
+          <button onClick={onSetEdit} className="resourceButton">
+            EDIT
+                    </button>
+          <button onClick={() => onDelete(resource.id)} className="resourceButton">
+            DELETE
+                    </button>
+        </div>
+      }
+    </div>
+  );
 }
 
-export default ResourceRow;
+export default ResourceRowRender;
